+++ lib-core/src/main/java/ch/hsr/adv/lib/core/logic/Builder.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.lib.core.logic;
 
 
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 
 /**
  * Builds a {@link ch.hsr.adv.lib.core.logic.domain.Session}
