+++ lib-core/src/main/java/ch/hsr/adv/lib/core/logic/CoreStringifyer.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.lib.core.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.core.access.JsonBuilderProvider;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.core.logic.domain.Session;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
