+++ lib-core/src/main/java/ch/hsr/adv/lib/core/logic/DefaultStringifyer.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.lib.core.logic;
 
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.core.access.GsonProvider;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.inject.Inject;
