+++ module-bb-array/src/main/java/ch/hsr/adv/lib/array/logic/ArrayBuilder.java
@@ -2,11 +2,11 @@ package ch.hsr.adv.lib.array.logic;
 
 import ch.hsr.adv.commons.array.logic.ConstantsArray;
 import ch.hsr.adv.commons.core.logic.domain.Module;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.lib.array.logic.domain.ArrayElement;
 import ch.hsr.adv.lib.core.logic.ADVModule;
 import ch.hsr.adv.lib.core.logic.Builder;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import com.google.inject.Singleton;
