+++ module-bb-array/src/main/java/ch/hsr/adv/lib/array/logic/ArrayStringifyer.java
@@ -2,9 +2,9 @@ package ch.hsr.adv.lib.array.logic;
 
 import ch.hsr.adv.commons.array.logic.ConstantsArray;
 import ch.hsr.adv.commons.core.logic.domain.Module;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.core.access.JsonBuilderProvider;
 import ch.hsr.adv.lib.core.logic.Stringifyer;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
