+++ module-bb-array/src/test/java/ch/hsr/adv/lib/array/logic/ArrayStringifyerTest.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.lib.array.logic;
 
 import ch.hsr.adv.commons.array.logic.ConstantsArray;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.core.logic.GuiceCoreModule;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import com.google.gson.JsonElement;
 import com.google.inject.Inject;
 import org.jukito.JukitoRunner;
