+++ module-bb-graph/src/main/java/ch/hsr/adv/lib/graph/logic/GraphBuilder.java
@@ -1,10 +1,10 @@
 package ch.hsr.adv.lib.graph.logic;
 
 import ch.hsr.adv.commons.core.logic.domain.Module;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.graph.logic.ConstantsGraph;
 import ch.hsr.adv.lib.core.logic.ADVModule;
 import ch.hsr.adv.lib.core.logic.Builder;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.graph.logic.domain.*;
 import com.google.inject.Singleton;
 import org.slf4j.Logger;
