+++ module-bb-stack/src/main/java/ch/hsr/adv/lib/stack/logic/StackStringifyer.java
@@ -1,10 +1,10 @@
 package ch.hsr.adv.lib.stack.logic;
 
 import ch.hsr.adv.commons.core.logic.domain.Module;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.commons.stack.logic.ConstantsStack;
 import ch.hsr.adv.lib.core.access.JsonBuilderProvider;
 import ch.hsr.adv.lib.core.logic.Stringifyer;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
