+++ module-bb-stack/src/test/java/ch/hsr/adv/lib/stack/logic/StackBuilderTest.java
@@ -1,7 +1,7 @@
 package ch.hsr.adv.lib.stack.logic;
 
 import ch.hsr.adv.commons.core.logic.domain.ADVElement;
-import ch.hsr.adv.lib.core.logic.domain.ModuleGroup;
+import ch.hsr.adv.commons.core.logic.domain.ModuleGroup;
 import ch.hsr.adv.lib.stack.logic.domain.StackTestModule;
 import ch.hsr.adv.lib.stack.logic.domain.TestConstants;
 import com.google.inject.Inject;
